Rework onboarding images
