+++ common/containers/OnboardModal/components/NotABankSlide.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import translate from 'translations';
 import OnboardSlide from './OnboardSlide';
-import onboardIconTwo from 'assets/images/onboarding_icon-02.svg';
+import onboardIconTwo from 'assets/images/onboarding/slide-02.svg';
 
 const NotABankSlide = () => {
   const header = translate('ONBOARD_bank_title');
