+++ common/containers/OnboardModal/components/SecureSlideThree.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import translate from 'translations';
 import OnboardSlide from './OnboardSlide';
-import onboardIconNine from 'assets/images/onboarding_icon-09.svg';
+import onboardIconNine from 'assets/images/onboarding/slide-09.svg';
 
 interface Props {
   site?: string;
