+++ common/containers/OnboardModal/components/index.ts
@@ -3,7 +3,7 @@ export { default as NotABankSlide } from './NotABankSlide';
 export { default as InterfaceSlide } from './InterfaceSlide';
 export { default as BlockchainSlide } from './BlockchainSlide';
 export { default as WhySlide } from './WhySlide';
-export { default as WhyMewSlide } from './WhyMewSlide';
+export { default as WhyMyCrypto } from './WhyMyCryptoSlide';
 export { default as SecureSlideOne } from './SecureSlideOne';
 export { default as SecureSlideTwo } from './SecureSlideTwo';
 export { default as SecureSlideThree } from './SecureSlideThree';
