+++ common/containers/OnboardModal/index.tsx
@@ -22,7 +22,7 @@ import {
   InterfaceSlide,
   BlockchainSlide,
   WhySlide,
-  WhyMewSlide,
+  WhyMyCrypto,
   SecureSlideOne,
   SecureSlideTwo,
   SecureSlideThree,
@@ -151,7 +151,7 @@ class OnboardModal extends React.Component<Props, State> {
       <InterfaceSlide key={3} />,
       <BlockchainSlide key={4} />,
       <WhySlide key={5} />,
-      <WhyMewSlide key={6} />,
+      <WhyMyCrypto key={6} />,
       <SecureSlideOne key={7} />,
       <SecureSlideTwo key={8} />,
       <SecureSlideThree key={9} />,
